requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
deprecation==2.1.0
Django==5.0.6
gotrue==2.5.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
load-dotenv==0.1.0
packaging==24.1
postgrest==0.16.8
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
realtime==1.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
WMI==1.5.1
